requirements.txt
Flask>=2.1.2
Flask-SQLAlchemy>=2.5.1
git+https://github.com/ssfdust/flask-sqlalchemy-stubs.git
mypy>=0.790
openai-whisper>=20230124
pyinstaller>=5.7.0
pywebview>=4.0.1
ruff>=0.0.210
sqlalchemy==1.4.39
